Hide tolerant parser behind --tolerant flag
The tolerant parser is useful, but can break parsing some seemingly-
well-formatted beta semvers. With that in mind, I'm making it a non-
default behavior for now. It would probably be worth digging into
this behavior in the parser to understand it better, but I don't
have the time to do it right now.